Change timing on pulsar timer
There was a bug in the pulsar timer gateware which works fine when under "self-pps" mode, but not under
actual pps signal. I have isolated this and changed so that everything now runs off the same "timing"
pps, which is muxed in precisely one place, no or gates elsewhere in the gateware.
This was manifest by offsets in the syncing of the channels coming out in the network packets. Hopefully
should correct once compiled.