Removed ability to select between Stokes and Phase.
Also removed the cordic ATAN and the Stokes I and V calculations which are unused anyway.
I know these are optimised away, but if I can save a bit of compiler time, that'll help as well.
As well as making the design less heavy for Simulink to process.

Same change effected in both WB and NB spectrometers, and a compile started. Will report results
in a github comment.