Added digital gain to both WB and NB spectrometers.
Had to requantise down to 18_17 again after the digital gain.

In WB spectrometer's case, I removed some unnecessary-looking delays
(I think they may have been associated with the ability to switch in
phase instead of Stokes), and packed the accumulators into a subsystem
to give Simulink a bit of an easier time.