Pulsartimer gateware packet timing fix
Fixes issue identified in https://skaafrica.atlassian.net/browse/AVN-299?filter=21338.
The problem was that the data_valid bits on the output of the buffer were not terribly intelligent,
and before the first sync pulse propagates through the system, it puts a few cycles worth of "valid"
data into the SPEAD packetiser, which then pushes a double header into the 10GbE block causing the
offset for the rest of time.

I put an enable-bit on the last delays coming through, so it waits for the first sync pulse before
it puts any d_v signal out.